requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pyperclip>=1.8.2
opencv-python>=4.8.0
numpy>=1.21.0
